Fix: cast per-attempt iteration caps to the solver-options width and size the analytical result array with a size type.
- Wrap max_iterations_per_attempt in static_cast<std::uint32_t> in the BOBYQA, L-BFGS-B, LM, augmented-Lagrangian, CMA-ES, and Newton-SQP IK solver adapters to match the seven sibling adapters that already do so, silencing -Wsign-conversion.
- Size analytical_types solutions array with static_cast<std::size_t>(MaxSolutions) to silence -Wsign-conversion on the array extent.

+++ lib/cartan-analytical/include/cartan/analytical/analytical_types.h
@@ -40,7 +40,7 @@ struct analytical_result
 {
     using position_type = Eigen::Vector<Scalar, N>;
 
-    std::array<position_type, MaxSolutions> solutions;
+    std::array<position_type, static_cast<std::size_t>(MaxSolutions)> solutions;
     int count{0};
 
     [[nodiscard]] auto begin() const { return solutions.begin(); }
